Refactored 'Create new application'
